pyproject.toml
[tool.poetry]
name = "maelstrom-api"
version = "1.0.0"
description = "The backend API for Maelstrom"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
asyncpg = "^0.22.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
aiofiles = "^0.6.0"
aiohttp = "^3.7.4"
python-dotenv = "^0.17.0"
Jinja2 = "^2.11.3"
itsdangerous = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
